Χρήση composable για κοινό χειρισμό συμβάντων
