+++ resources/ts/components/frontend/form/field/Date.vue
@@ -27,5 +27,6 @@ const formStore = useFormStore();
         :disabled="disabled"
         :required="field.required ? 'true' : undefined"
         v-model="formStore.field[props.field.id]"
+        autocomplete="off"
     />
 </template>
